Refactor server sources organization
