+++ server/src/infrastructure/middlewares/auth.js
@@ -1,5 +1,5 @@
 const jwt = require('jsonwebtoken')
-const config = require('../config')
+const config = require('../../config')
 
 function _extractAccessTokenFromAuthorizationHeader (authorizationHeader) {
   let accessToken
